+++ genomepy/annotation/__init__.py
@@ -140,6 +140,28 @@ class Annotation:
             self.genome_contigs = None  # noqa
         super(Annotation, self).__setattr__(name, value)
 
+    def attributes(self, annot: Union[str, pd.DataFrame] = "gtf"):
+        """
+        list all attributes present in the GTF attribute field.
+
+        Parameters
+        ----------
+        annot : str or pd.Dataframe, optional
+            any GTF in dataframe format, or the default GTF.
+
+        Returns
+        -------
+        list
+            with attributes
+        """
+        df = _parse_annot(self, annot)
+        attributes = set()
+        for feature in df["feature"].unique():
+            f_attributes = df[df["feature"] == feature]["attribute"].head(1).values[0]
+            f_attributes = re.findall(r'\s*(.+?)\s*".+?"\s*;', f_attributes)
+            attributes.update(f_attributes)
+        return sorted(attributes)
+
     def from_attributes(
         self, field, annot: Union[str, pd.DataFrame] = "gtf", check=True
     ):